Organize DLT modules by protocol version and remove legacy files
