+++ src/dlt/mod.rs
@@ -1,5 +1,3 @@
-mod dlt_common;
-mod dlt_protocol;
 pub mod v1;
 
 pub mod error;
